tests: change dataset_name to a tuple
